+++ lca_algebraic/lca.py
@@ -3,21 +3,30 @@ import re
 from collections import OrderedDict
 from dataclasses import dataclass
 from types import FunctionType
-from typing import Dict, List, Union
+from typing import Dict, List, Optional, Tuple, Union
 
 import brightway2 as bw
 import numpy as np
 import pandas as pd
+from pandas import DataFrame
 from peewee import DoesNotExist
-from sympy import Basic, Symbol, lambdify, parse_expr, simplify, symbols
+from pint import Quantity, Unit
+from sympy import Basic, Expr, Symbol, lambdify, parse_expr, simplify, symbols
 from sympy.printing.numpy import NumPyPrinter
+from typing_extensions import deprecated
 
-from . import ValueOrExpression
-from .activity import ActivityExtended, DbContext, newActivity, with_db_context
-from .axis_dict import AxisDict
-from .base_utils import TabbedDataframe, _actName, _getDb, _user_functions
+from . import newActivity
+from .activity import ActivityExtended
+from .axis_dict import NO_AXIS_NAME, AxisDict
+from .base_utils import (
+    TabbedDataframe,
+    ValueOrExpression,
+    _actName,
+    _getDb,
+    _user_functions,
+)
 from .cache import ExprCache, LCIACache
-from .database import BIOSPHERE_PREFIX, _isForeground
+from .database import BIOSPHERE_PREFIX, DbContext, _isForeground, with_db_context
 from .log import logger, warn
 from .methods import method_name, method_unit
 from .params import (
@@ -145,24 +154,17 @@ def _replace_symbols_with_params_in_exp(expr: Basic):
     return expr
 
 
-def _modelToExpr(model: ActivityExtended, methods, alpha=1, axis=None):
-    """
-    Compute expressions corresponding to a model for each impact, replacing activities by the value of its impact
-
-    Return
-    ------
-    <list of expressions (one per impact)>, <list of required param names>
-
-    """
-
+def _cachedActToExpression(
+    model: ActivityExtended, alpha=1.0, axis=None, for_inventory=False
+) -> Tuple[Expr, Dict[Symbol, ActivityExtended]]:
     # Try to load from cache
     with ExprCache() as cache:
-        key = (model, axis)
+        key = (model, axis, for_inventory)
 
         if key not in cache.data:
             logger.debug(f"{model} was not in exrepssion cache, computing...")
 
-            expr, actBySymbolName = actToExpression(model, axis=axis)
+            expr, actBySymbolName = actToExpression(model, axis=axis, for_inventory=for_inventory)
 
             cache.data[key] = (expr, actBySymbolName)
         else:
@@ -170,12 +172,9 @@ def _modelToExpr(model: ActivityExtended, methods, alpha=1, axis=None):
 
         expr, actBySymbolName = cache.data[key]
 
-    logger.debug("Alpha passed %s", alpha)
-
-    # logger.debug("Raw expression for %s/%s : '%s'", model, str(methods), expr)
-    # logger.debug("Act by symbol : %s", actBySymbolName)
-    # if logger.isEnabledFor("DEBUG") :
-    # logger.debug(f"Length of expression : {len(str(expr))}")
+    # Using units ? Only keep the magnitude, drop the units
+    if isinstance(alpha, Quantity):
+        alpha = alpha.magnitude
 
     expr = expr * alpha
 
@@ -186,14 +185,29 @@ def _modelToExpr(model: ActivityExtended, methods, alpha=1, axis=None):
     for act_name, act in actBySymbolName.items():
         pureTechActBySymbol[act_name] = _createTechProxyForBio(act, model.key[0])
 
+    return expr, pureTechActBySymbol
+
+
+def _modelToExpr(model: ActivityExtended, methods, alpha=1, axis=None):
+    """
+    Compute expressions corresponding to a model for each impact, replacing activities by the value of its impact
+
+    Return
+    ------
+    <list of expressions (one per impact)>, <list of required param names>
+
+    """
+
+    expr, act_by_symbol = _cachedActToExpression(model, alpha=alpha, axis=axis)
+
     # Compute LCA for background activities
-    lcas = _multiLCAWithCache(pureTechActBySymbol.values(), methods)
+    lcas = _multiLCAWithCache(act_by_symbol.values(), methods)
 
     # For each method, compute an algebric expression with activities replaced by their values
     exprs = []
     for method in methods:
         # Replace activities by their value in expression for this method
-        sub = dict({symbol: lcas[(act, method)] for symbol, act in pureTechActBySymbol.items()})
+        sub = dict({symbol: lcas[(act, method)] for symbol, act in act_by_symbol.items()})
 
         expr_curr = expr.xreplace(sub)
 
@@ -247,11 +261,6 @@ def _lambdify(expr: Basic, expanded_params):
         return static_func
 
 
-def compute_param_formulas(param_values: Dict, required_params: List[str]):
-    # Compute the values params computed from formulas (and not specified in input)
-    pass
-
-
 @dataclass
 class ValueContext:
     """Represents a result value, with all parameters values used in context"""
@@ -268,14 +277,14 @@ class LambdaWithParamNames:
 
     _use_sympy_cse = False
 
-    def __init__(self, exprOrDict, expanded_params=None, params=None, sobols=None):
+    def __init__(self, expr: Expr, expanded_params=None, params=None, sobols=None):
         """Computes a lamdda function from expression and list of expected parameters.
         you can provide either the list pf expanded parameters (full vars for enums) for the 'user' param names
         """
 
-        if isinstance(exprOrDict, dict):
+        if isinstance(expr, dict):
             # Come from JSON serialization
-            obj = exprOrDict
+            obj = expr
             # LIst of required params for this lambda
             self.params: List[str] = obj["params"]
 
@@ -287,12 +296,12 @@ class LambdaWithParamNames:
             self.sobols = obj["sobols"]
 
         else:
-            self.expr = exprOrDict
+            self.expr = expr
             self.params = params
 
             if expanded_params is None:
                 if params is None:
-                    expanded_params = _free_symbols(exprOrDict)
+                    expanded_params = _free_symbols(expr)
                     params = _expanded_names_to_names(expanded_params)
                     self.params = params
 
@@ -304,7 +313,7 @@ class LambdaWithParamNames:
             elif self.params is None:
                 self.params = _expanded_names_to_names(expanded_params)
 
-            self.lambd = _lambdify(exprOrDict, expanded_params)
+            self.lambd = _lambdify(expr, expanded_params)
             self.expanded_params = expanded_params
             self.sobols = sobols
 
@@ -354,13 +363,14 @@ def lambdify_expr(expr):
     return LambdaWithParamNames(expr, params=[param.name for param in _param_registry().values()])
 
 
-def _preMultiLCAAlgebric(model: ActivityExtended, methods, alpha=1, axis=None):
+def _preMultiLCAAlgebric(model: ActivityExtended, methods, alpha: ValueOrExpression = 1, axis=None):
     """
     This method transforms an activity into a set of functions ready to compute LCA very fast on a set on methods.
     You may use is and pass the result to postMultiLCAAlgebric for fast computation on a model that does not change.
 
     This method is used by multiLCAAlgebric
     """
+
     with DbContext(model):
         exprs = _modelToExpr(model, methods, alpha=alpha, axis=axis)
 
@@ -380,7 +390,7 @@ class ResultsWithParams:
     params: Dict
 
 
-def _postMultiLCAAlgebric(methods, lambdas: List[LambdaWithParamNames], with_params=False, **params):
+def _postMultiLCAAlgebric(methods, lambdas: List[LambdaWithParamNames], with_params=False, unit: Unit = None, **params):
     """
     Compute LCA for a given set of parameters and pre-compiled lambda functions.
     This function is used by **multiLCAAlgebric**
@@ -437,7 +447,7 @@ def _postMultiLCAAlgebric(methods, lambdas: List[LambdaWithParamNames], with_par
 
     result = pd.DataFrame(
         res,
-        index=[method_name(method) + "[%s]" % method_unit(method) for method in methods],
+        index=[method_name(method) + "[%s]" % method_unit(method, fu_unit=unit) for method in methods],
     ).transpose()
 
     if with_params:
@@ -477,9 +487,10 @@ def compute_value(formula, **params):
     return value_context.value
 
 
+@deprecated("multiLCAAlgebric is deprecated, use compute_impacts instead")
 def multiLCAAlgebric(*args, **kwargs):
     """deprecated. `compute_impacts()` instead"""
-    logger.warn("multiLCAAlgebric is deprecated, use compute_impacts instead")
+    warn("multiLCAAlgebric is deprecated, use compute_impacts instead")
     return compute_impacts(*args, **kwargs)
 
 
@@ -520,14 +531,85 @@ def _params_dataframe(param_values: Dict[str, float]):
 SingleOrMultipleFloat = Union[float, List[float], np.ndarray]
 
 
+def compute_inventory(
+    model: ActivityExtended, functional_unit=1, as_dict=False, fields=["database", "name", "location", "unit"], **params
+):
+    """
+
+    This method computes the inventory of background activities for a given scenario / values of parameters.
+
+    Parameters
+    ----------
+    model:
+        Root activity
+
+    functional_unit:
+        Quantitity to divide the inventory. 1 by default
+
+    as_dict:
+        If true, returns a dict of act => value. If false (default) returns a dataframe
+
+    fields:
+        List of fields to be added in the ouput dataframe
+
+    params:
+        All other attributes are treated as values of lca_algebraic parameters.
+        If not specified, each parameters takes its default value.
+
+    Returns
+    -------
+    Dataframe or Dict of act => value
+
+    """
+
+    expr, act_by_symbol = _cachedActToExpression(model, alpha=1 / functional_unit, for_inventory=True)
+
+    val_by_act_name = compute_value(expr, **params)
+
+    def get_act_by_name(act_name):
+        """Get activitiy. Possibly derefernce it (if is a bio proxy)"""
+        act = act_by_symbol[symbols(act_name)]
+        if "isProxy" in act:
+            exchanges = act.listExchanges()
+            assert len(exchanges) == 1
+            name, act, amount = exchanges[0]
+        return act
+
+    # Transform to dict of act => value
+    val_by_act = dict()
+    for act_name, value in val_by_act_name.items():
+        if act_name == NO_AXIS_NAME:
+            continue
+
+        act = get_act_by_name(act_name)
+        val_by_act[act] = value
+
+    if as_dict:
+        return val_by_act
+
+    # Transform to dataframe
+    items = []
+    for act, value in val_by_act.items():
+        item = dict()
+
+        for field in fields:
+            if field in act:
+                item[field] = act[field]
+
+        item["value"] = value
+        items.append(item)
+
+    return DataFrame(items)
+
+
 def compute_impacts(
     models,
     methods,
     axis=None,
-    functional_unit: ValueOrExpression = 1,
-    return_params: bool = False,
-    description: str = None,
-    **params: Dict[str, SingleOrMultipleFloat],
+    functional_unit=1,
+    return_params=False,
+    description=None,
+    **params,
 ):
     """
     Main parametric LCIA method :
@@ -607,6 +689,10 @@ def compute_impacts(
     # Gather all param values (even default and computed)
     params_all = dict()
 
+    # Single method provided ?
+    if isinstance(methods, tuple):
+        methods = [methods]
+
     for model, alpha in models.items():
         if type(model) is tuple:
             model, alpha = model
@@ -625,7 +711,10 @@ def compute_impacts(
 
             lambdas = _preMultiLCAAlgebric(model, methods, alpha=alpha, axis=axis)
 
-            res = _postMultiLCAAlgebric(methods, lambdas, with_params=return_params, **params)
+            unit: Optional[Unit] = functional_unit.units if isinstance(functional_unit, Quantity) else None
+
+            res = _postMultiLCAAlgebric(methods, lambdas, with_params=return_params, unit=unit, **params)
+
             if return_params:
                 df = res.dataframe
                 params_all.update(res.params)
@@ -709,7 +798,7 @@ def _createTechProxyForBio(act_key, target_db):
 
             # Create biosphere proxy in User Db
             res = newActivity(
-                target_db, name, act["unit"], {act: 1}, code=code_to_find, isProxy=True
+                target_db, name, act["unit"], {act: 1}, code=code_to_find, switchActivity=True, isProxy=True
             )  # add a this flag to distinguish this dummy activity from others
             return res
     else:
@@ -756,7 +845,7 @@ def _tag_expr(expr, act, axis):
 
 
 @with_db_context(arg="act")
-def actToExpression(act: ActivityExtended, axis=None):
+def actToExpression(act: ActivityExtended, axis=None, for_inventory=False):
     """Computes a symbolic expression of the model, referencing background activities and model parameters as symbols
 
     Returns
@@ -775,7 +864,7 @@ def actToExpression(act: ActivityExtended, axis=None):
         if not (db_name, code) in act_symbols:
             act = _getDb(db_name).get(code)
             name = act["name"]
-            base_slug = _slugify(name)
+            base_slug = _slugify("_" + name)
 
             slug = base_slug
             i = 1
@@ -785,7 +874,13 @@ def actToExpression(act: ActivityExtended, axis=None):
 
             act_symbols[(db_name, code)] = symbols(slug)
 
-        return act_symbols[(db_name, code)]
+        act_symbol = act_symbols[(db_name, code)]
+
+        if for_inventory:
+            # When computing inventories, we build Dict of {activity_id => amount_or_formula}
+            return AxisDict({str(act_symbol): 1})
+        else:
+            return act_symbol
 
     # Local cache of expressions
 
@@ -842,6 +937,7 @@ def actToExpression(act: ActivityExtended, axis=None):
 
     if isinstance(expr, float):
         expr = simplify(expr)
+
     else:
         # Replace fixed params with their default value
         expr = _replace_fixed_params(expr, _fixed_params().values())