+++ lca_algebraic/stats.py
@@ -3,7 +3,7 @@ import random
 import warnings
 from collections import defaultdict
 from time import time
-from typing import List, Tuple, Type
+from typing import Dict, List, Tuple, Type
 
 import numpy as np
 import seaborn as sns
@@ -30,11 +30,14 @@ from sympy import (
 )
 from sympy.core.operations import AssocOp
 
-from . import ValueOrExpression, method_name
-from .base_utils import _display_tabs, displayWithExportButton, r_squared
+from .base_utils import (
+    ValueOrExpression,
+    _display_tabs,
+    displayWithExportButton,
+    r_squared,
+)
 from .database import DbContext, with_db_context
 from .lca import (
-    Dict,
     LambdaWithParamNames,
     Symbol,
     _expanded_names_to_names,
@@ -49,7 +52,7 @@ from .lca import (
     pd,
 )
 from .log import warn
-from .methods import method_unit
+from .methods import method_name, method_unit
 from .params import (
     FixedParamMode,
     NameType,
@@ -1295,11 +1298,11 @@ def compare_simplified(
         params, _ = _generate_random_params(100000, sample_method=StochasticMethod.RAND)
 
         # Run  Monte Carlo on full model
-        Y1 = _compute_stochastics([lambd], [method], params=params)
+        Y1 = _compute_stochastics([lambd], [method], functional_unit=functional_unit, params=params)
         d1 = Y1[Y1.columns[0]]
 
         # Run monte carlo of simplified model
-        Y2 = _compute_stochastics([simpl_lambd], [method], params=params)
+        Y2 = _compute_stochastics([simpl_lambd], [method], functional_unit=functional_unit, params=params)
         d2 = Y2[Y2.columns[0]]
 
         r_value = r_squared(Y1, Y2)