+++ RELEASE_NOTES.md
@@ -1,12 +1,17 @@
-# 
-* Fix bug in compare_simplified introduced since the release of *functional_unit* 
+# 1.2 
 
+* Fix bug in compare_simplified introduced since the release of *functional_unit*
+* Removed pypardisio from iOS/ARM build
+* Fix major bug with axis being in different order when used with multiple methods 
+* Added `compute_inventory()`
+* Added support for units  with Pint : automatic check of physical units 
 
 # 1.1.2
 
 Bugfix release :
 * Fix major bug #58 with axis being in different order when used with multiple methods
 * Fix bug #54 : unexpected behaviour with '*' wildcard and 'old_amount' parameter
+
 
 # 1.1